Backport FFI-based File fixes to defer load
See #6760 and related commits to these files on master.
